Move scheduler files to proper folders
